@@ -1,15 +1,15 @@
-const custom = require('../webpack.config.js');
-
 module.exports = {
   "stories": [
     "../src/**/*.stories.@(js|jsx|ts|tsx)"
   ],
   "addons": [
-    '@storybook/addon-links',
-    '@storybook/addon-essentials',
-    '@storybook/present-create-react-app'
+    "@storybook/addon-links",
+    "@storybook/addon-essentials",
+    "@storybook/addon-interactions",
+    "@storybook/preset-create-react-app"
   ],
-  webpackFinal: async (config) => {
-    return { ...config, module: { ...config.module, rules: custom.module.rules } };
-  },
+  "framework": "@storybook/react",
+  "core": {
+    "builder": "@storybook/builder-webpack5"
+  }
 }
